Extract agent log forwarder
